@@ -2,7 +2,7 @@
  * 认证服务类
  */
 import {Injectable} from '@angular/core';
-import {Subject} from 'rxjs/Subject';
+import {Subject} from 'rxjs';
 
 @Injectable()
 export class AuthService {
